refactor(sdk): move the iam placeholder helpers into their own module
The placeholder spelling, the name rule it forces, and the token map the
callbacks read are one concern with a backend contract behind it, so they
get a named module next to `network`/`mcp`/`signature` rather than more
lines in sandboxApi. `utils` is the other candidate, but that is where
generic mechanism lives (runtime detection, shellQuote,
class_method_variant), not domain knowledge.

sandboxApi keeps what is network-shaped: building the transform context
around the map, and checking names as tokens are registered.
